Give Reviewer posting its own routes module
routes/github.py is past 7k lines and keeps absorbing unrelated
features. The reviewer grant/post surface is self-contained, so move
its bodies, helpers, and routes into routes/reviewer.py under the
same /api/github prefix; the exactly-once claim machinery stays in
app.reviewer_automation. Tests now patch the helpers at their new
home. Also document validated_comment's deliberately coarse HTML
guard (List<T> and friends are rejected along with real tags).
