+++ backend/app/main.py
@@ -79,6 +79,7 @@ from app.routes import (
   contribution_reviews_router,
   chat_waits_router,
   debug_router, delegations_router, fs_router, goal_plans_router, github_router,
+  reviewer_router,
   identity_router,
   owner_approvals_router,
   media_router,
@@ -868,6 +869,7 @@ app.include_router(uploads_router)
 app.include_router(media_router)
 app.include_router(secrets_router)
 app.include_router(github_router)
+app.include_router(reviewer_router)
 app.include_router(contribution_reviews_router)
 app.include_router(identity_router)
 app.include_router(push_router)
